backend/requirements.txt
blinker==1.9.0
click==8.2.1
Flask==3.1.1
flask-cors==6.0.1
Flask-JWT-Extended==4.6.0
Flask-SQLAlchemy==3.1.1
psycopg2-binary==2.9.10
python-dotenv==1.0.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
numpy==2.3.1
packaging==25.0
pandas==2.3.1
patsy==1.0.1
python-dateutil==2.9.0.post0
pytz==2025.2
scipy==1.16.0
six==1.17.0
statsmodels==0.14.5
tzdata==2025.2
Werkzeug==3.1.3
